+++ management-ui/src/components/Home.jsx
@@ -1,5 +1,6 @@
 import React, { Component } from 'react';
 import { withStyles } from '@material-ui/core';
+import { OpenInNew } from '@material-ui/icons';
 
 const styles = theme => ({
   mainContent: {
@@ -19,8 +20,25 @@ class Home extends Component {
           </h1>
           <p>
             <small>
-              Copyright © 2018 Henrik Franciscus Leppä<br />
-              All rights reserved.
+              Copyright © 2018&ndash;2019 Henrik Franciscus Leppä
+            </small>
+          </p>
+          <p>
+            <small>
+              This is free software. You may redistribute copies of it under the
+              terms of the
+              {' '}
+              <a
+                href="https://opensource.org/licenses/MIT"
+                target="_blank"
+                rel="external noopener noreferrer"
+              >
+                <abbr title="Opens in a new tab/window">
+                  <OpenInNew fontSize="inherit" />
+                </abbr>
+                MIT License
+              </a>
+              . There is NO WARRANTY, to the extent permitted by law.
             </small>
           </p>
         </header>