+++ CHANGELOG.md
@@ -21,6 +21,10 @@ Versioning].
 - Node.js & npm versions to `package.json` files
 - `.gitattributes` file
 
+### Changed
+
+- From unlicensed to MIT License
+
 ### Security
 
 - Rebuild code with Create React App 3.0.1, and update all dependencies